refactor(domain):add common action to abstract class
