+++ src/main/java/com/game/domain/RemoteChessPlayer.java
@@ -1,11 +1,13 @@
 package com.game.domain;
 
+
+import com.game.service.NetworkService;
+
 /**
  * @author VIRIYA
  * @create 2020/10/24 2:15
  */
 public class RemoteChessPlayer extends AbstractPlayer {
-
 
     /**
      * message of player send to other player
@@ -16,13 +18,6 @@ public class RemoteChessPlayer extends AbstractPlayer {
      */
     @Override
     public Point play() {
-        // get point by network
-
-        return null;
-    }
-
-    @Override
-    public Result gameResult() {
         return null;
     }
 
@@ -30,11 +25,5 @@ public class RemoteChessPlayer extends AbstractPlayer {
     public Point startPlay() {
         return null;
     }
-
-    @Override
-    public boolean checkEmptySlot(Point point) {
-        return false;
-    }
-
 
 }
